% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{aarrow_dependency}
\alias{aarrow_dependency}
\alias{a_arrow}
\title{Functions}
\usage{
aarrow_dependency()

a_arrow(arrow = NA, ...)
}
\arguments{
\item{arrow}{Arrow settings.}

\item{...}{Any \code{aframer} option of element.}
}
\description{
\code{aarrow} dependency.
}
\examples{
library(aframer)

a_scene(
  a_dependency(),
  aarrow_dependency(),
  a_arrow(arrow = aframe_opts(color = "blue"))
)

}
